Revert "Incorprated a patch from Tuxi on the #iodoom3 irc channel."
This reverts commit e9f219d2122874974d1271b4713c9fef06c22a2d.

+++ neo/ui/SimpleWindow.cpp
@@ -254,35 +254,35 @@ void idSimpleWindow::Redraw(float x, float y) {
 	textRect.Offset(-x, -y);
 }
 
-long idSimpleWindow::GetWinVarOffset( idWinVar *wv, drawWin_t* owner) {
-	long ret = -1;
+int idSimpleWindow::GetWinVarOffset( idWinVar *wv, drawWin_t* owner) {
+	int ret = -1;
 
 	if ( wv == &rect ) {
-		ret = (long)&( ( idSimpleWindow * ) 0 )->rect;
+		ret = (int)&( ( idSimpleWindow * ) 0 )->rect;
 	}
 
 	if ( wv == &backColor ) {
-		ret = (long)&( ( idSimpleWindow * ) 0 )->backColor;
+		ret = (int)&( ( idSimpleWindow * ) 0 )->backColor;
 	}
 
 	if ( wv == &matColor ) {
-		ret = (long)&( ( idSimpleWindow * ) 0 )->matColor;
+		ret = (int)&( ( idSimpleWindow * ) 0 )->matColor;
 	}
 
 	if ( wv == &foreColor ) {
-		ret = (long)&( ( idSimpleWindow * ) 0 )->foreColor;
+		ret = (int)&( ( idSimpleWindow * ) 0 )->foreColor;
 	}
 
 	if ( wv == &borderColor ) {
-		ret = (long)&( ( idSimpleWindow * ) 0 )->borderColor;
+		ret = (int)&( ( idSimpleWindow * ) 0 )->borderColor;
 	}
 
 	if ( wv == &textScale ) {
-		ret = (long)&( ( idSimpleWindow * ) 0 )->textScale;
+		ret = (int)&( ( idSimpleWindow * ) 0 )->textScale;
 	}
 
 	if ( wv == &rotate ) {
-		ret = (long)&( ( idSimpleWindow * ) 0 )->rotate;
+		ret = (int)&( ( idSimpleWindow * ) 0 )->rotate;
 	}
 
 	if ( ret != -1 ) {
